feat(useTransition): rework `expires` as a number > 0
When `expires` is > 0, it delays the unmounting of an item for a number of milliseconds. Once the timeout ends, the item will try to unmount itself. If the caller re-renders before then, the item will stay mounted, since it technically has no expired yet.

Previously, when the timeout ended, it would force the caller to re-render no matter what. I've decided that behavior is much less useful than delayed expiration.